Refactor repeated code in optimisation._nlopt
